backend/requirements.txt
fastapi>=0.103,<0.110
uvicorn[standard]>=0.23,<0.29
python-dotenv>=1.0,<1.1
pydantic>=1.10,<2.0
requests>=2.31,<3.0
google-auth>=2.29,<3.0
